Unify Invalid, Degenerate and Twist suites on standalone.
